refactor: group similar images together
