+++ package.json
@@ -20,7 +20,8 @@
     "next-themes": "^0.2.1",
     "react": "^18",
     "react-dom": "^18",
-    "react-icons": "^4.12.0"
+    "react-icons": "^4.12.0",
+    "sharp": "^0.33.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^18.4.3",
